[refactor] #25 service controller 패키지에서 이동 service 패키지로 이동
