+++ src/main/java/com/ggang/be/domain/school/application/SchoolServiceImpl.java
@@ -1,6 +1,6 @@
 package com.ggang.be.domain.school.application;
 
-import com.ggang.be.api.school.controller.SchoolService;
+import com.ggang.be.api.school.service.SchoolService;
 import com.ggang.be.domain.school.SchoolEntity;
 import com.ggang.be.domain.school.dto.SchoolSearchVo;
 import com.ggang.be.domain.school.infra.SchoolRepository;
